fix(console): icon-rail collapse + relocate sidebar store
Addresses PR #417 review:
- Move sidebar-store from shell/ to app/, alongside session.ts — the
  persisted-app-shell-state precedent; its only consumer is root-layout.
- Trim the store's design-rationale comment; rationale lives in the PR
  body per code-comments.md.
- Set <Sidebar collapsible="icon"> so a persisted collapse shows a usable
  icon rail rather than a fully hidden (offcanvas) sidebar.

Verified at runtime: collapse → icon rail (icons visible) → reload → state persists.
